+++ src/main/java/codesquad/web/ProductController.java
@@ -1,0 +1,26 @@
+package codesquad.web;
+
+import codesquad.service.ProductService;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Controller;
+import org.springframework.ui.Model;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+
+@Controller
+public class ProductController {
+    @Autowired
+    private ProductService productService;
+
+    @GetMapping("/products")
+    public String products(Model model) {
+        model.addAttribute("products", productService.findAll());
+        return "products";
+    }
+
+    @GetMapping("/product/{id}")
+    public String product(@PathVariable long id, Model model) {
+        model.addAttribute("product", productService.findById(id));
+        return "product";
+    }
+}
